add new metrics

+++ dto/info_dto.go
@@ -1,5 +1,5 @@
 package dto
 
 type CountResponse struct {
-	Total int `json:"total"`
+	Total float64 `json:"total"`
 }
